+++ ruby/test/mainapp_test.rb
@@ -201,7 +201,7 @@ END
 Usage: #{File.basename($0)} [<options>] [<file-or-directory>...]
   -h, --help             : show help
       --version          : print version
-  -s <STYLE>             : report style (verbose/simple/plain/quiet, or v/s/p/q)
+  -s <REPORT-STYLE>      : verbose/simple/compact/plain/quiet, or v/s/c/p/q
   -F <PATTERN>           : filter topic or spec with pattern (see below)
       --color[={on|off}] : enable/disable output coloring forcedly
   -C, --create           : print test code skeleton
@@ -251,6 +251,7 @@ END
 
     it "[!0qd92] '-s verbose' or '-sv' option prints test results in verbose mode." do
       expected = <<END
+## _tmp_test.rb
 * <b>Parent</b>
   * <b>Child1</b>
     - [<B>pass</B>] 1+1 should be 2
@@ -272,9 +273,12 @@ END
       assert_eq serr, ""
     end
 
-    it "[!ef5v7] '-s simple' or '-ss' option prints test results in simple mode." do
+    it "[!zfdr5] '-s simple' or '-ss' option prints test results in simple mode." do
       expected = <<END
-#{@testfile}: <B>.</B><B>.</B><R>f</R><R>E</R><Y>s</Y><Y>t</Y><B>.</B><B>.</B>
+## _tmp_test.rb
+* <b>Parent</b>: <B>.</B><B>.</B>
+  * <b>Child1</b>: <B>.</B><B>.</B>
+  * <b>Child2</b>: <R>f</R><R>E</R>
 ----------------------------------------------------------------------
 END
       #
@@ -284,6 +288,23 @@ END
       assert_eq serr, ""
       #
       ret, sout, serr = run("-s", "simple", @testfile)
+      assert_eq ret, 2
+      assert edit_actual(sout).start_with?(edit_expected(expected)), "invalid testcase output"
+      assert_eq serr, ""
+    end
+
+    it "[!ef5v7] '-s compact' or '-sc' option prints test results in compact mode." do
+      expected = <<END
+#{@testfile}: <B>.</B><B>.</B><R>f</R><R>E</R><Y>s</Y><Y>t</Y><B>.</B><B>.</B>
+----------------------------------------------------------------------
+END
+      #
+      ret, sout, serr = run("-sc", @testfile)
+      assert_eq ret, 2
+      assert edit_actual(sout).start_with?(edit_expected(expected)), "invalid testcase output"
+      assert_eq serr, ""
+      #
+      ret, sout, serr = run("-s", "compact", @testfile)
       assert_eq ret, 2
       assert edit_actual(sout).start_with?(edit_expected(expected)), "invalid testcase output"
       assert_eq serr, ""
@@ -325,6 +346,7 @@ END
 
     it "[!yz7g5] '-F topic=...' option filters topics." do
       expected = <<END
+## _tmp_test.rb
 * <b>Parent</b>
   * <b>Child1</b>
     - [<B>pass</B>] 1+1 should be 2
@@ -340,6 +362,7 @@ END
 
     it "[!ww2mp] '-F spec=...' option filters specs." do
       expected = <<END
+## _tmp_test.rb
 * <b>Parent</b>
   * <b>Child1</b>
     - [<B>pass</B>] 1-1 should be 0
@@ -354,6 +377,7 @@ END
 
     it "[!8uvib] '-F tag=...' option filters by tag name." do
       expected = <<'END'
+## _tmp_test.rb
 * <b>Parent</b>
   * <b>Child1</b>
     - [<B>pass</B>] 1-1 should be 0
@@ -373,6 +397,7 @@ END
 
     it "[!m0iwm] '-F sid=...' option filters by spec id." do
       expected = <<'END'
+## _tmp_test.rb
 * <b>Parent</b>
   - <b>When x is negative</b>
     - [<B>pass</B>] [!6hs1j] x*x is positive.
@@ -387,6 +412,7 @@ END
 
     it "[!noi8i] '-F' option supports negative filter." do
       expected = <<'END'
+## _tmp_test.rb
 * <b>Parent</b>
   * <b>Child1</b>
     - [<B>pass</B>] 1+1 should be 2